--- a/Gemfile
+++ b/Gemfile
@@ -1,4 +1,0 @@
-source 'http://rubygems.org'
-gem 'sass', '3.3.9'
-gem 'compass', '1.0.0.alpha.19'
-gem 'sass-globbing', '1.1.0'