--- a/Gruntfile.js
+++ b/Gruntfile.js
@@ -14,9 +14,7 @@ module.exports = function(grunt) {
     require('time-grunt')(grunt);
 
     // Loads grunt tasks found on package.json automatically
-    require('load-grunt-tasks')(grunt, {
-        pattern: 'grunt-*'
-    });
+    require('jit-grunt')(grunt);
 
     // Requiring our custom grunt tasks from the toolset
     // TODO: move to npm
@@ -42,6 +40,7 @@ module.exports = function(grunt) {
 
     // Loads grunt tasks' configuration in lib/grunt-tasks/<task-name>.js
     require('load-grunt-config')(grunt, {
+        jitGrunt: true,
         data: {
             config: config
         },