--- a/app/css/main.css.scss
+++ b/app/css/main.css.scss
@@ -12,7 +12,10 @@
     "1.base",
     "2.layout",
     "3.helpers",
-    "../components/**/*",
+    "../components/button/button",
+    "../components/docs-skeleton/docs-skeleton",
+    "../components/hello-world/hello-world",
+    "../components/tweet/tweet",
     "4.print",
     "5.states",
     "6.shame";