--- a/lib/grunt-tasks/watch.js
+++ b/lib/grunt-tasks/watch.js
@@ -8,7 +8,7 @@ module.exports = {
             '<%= config.srcDir %>' + '/**/*.scss',
             '<%= config.libDir %>/toolset/grunt-tasks/assembler/**/*.scss'
         ],
-        tasks: ['compass', 'autoprefixer']
+        tasks: ['sass', 'autoprefixer']
     },
     jade: {
         files: ['<%= config.srcDir %>' + '/**/*.jade'],
@@ -28,4 +28,3 @@ module.exports = {
       tasks: ['styleguideAssembler', 'stylesAssembler']
     }
 };
-