@idflood 2015-04-17
I made some test with the current Basler project. jit-grunt + libsass definitely make things much faster, here are some number taken from the watch task:

### Jade:
#### before

    loading tasks       2.7s  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 54%
    pagesAssembler:dev  2.3s  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 46%
    Total 5s

#### after

    loading tasks       443ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 15%
    pagesAssembler:dev   2.4s  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 85%
    Total 2.9s


### Sass:
#### before

    loading tasks      2.7s  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 61%
    sass:dev           1.5s  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 35%
    autoprefixer:all  147ms  ▇▇▇▇ 3%
    Total 4.4s

#### after

    loading tasks               444ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 55%
    loading grunt-sass           14ms  ▇▇ 2%
    sass:dev                     74ms  ▇▇▇▇▇▇▇▇▇▇ 9%
    loading grunt-autoprefixer  135ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 17%
    autoprefixer:all            137ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 17%
    Total 805ms