@idflood 2015-04-17
When switching existing project you may have to:

- run `npm install`
- delete the `node_modules/grunt-contrib-sass` folder since it will be loaded by default, even if `grunt-sass` is present.