@lucalanca 2015-04-17
Excelent :+1: 

Just as a note, this PR also includes grunt-jit, not only switching from ruby-sass to libsass.