@lejoe 2015-04-24
Merged directly to master as the styleguide branch is now there